fix: nested FBX node not found in Appender and GeneratedMeshAssetSO
Resolved an issue where nested FBX nodes were not properly found in BlendShapeAppender
when generating the FBX first and using it to copy to the mesh asset, causing failures
when the Unity mesh was not an exact match.

Also fixed a related issue in GeneratedMeshAssetSO.ApplyMesh(), where the mesh application
process did not handle nested nodes correctly.